% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{download_model}
\alias{download_model}
\title{Download the model}
\usage{
download_model(path, max.time = 9999)
}
\arguments{
\item{path}{the path to store the model, make sure ended with .rds}

\item{max.time}{the max time (seconds) for downloading the model file. The default is 9999 seconds.}
}
\value{
No return value
}
\description{
Download the model
}
\examples{
download_model(paste0(getwd(), "theModel.rds"))
}
\author{
Shangchen Song
}
