% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{load_model}
\alias{load_model}
\title{Load the model}
\usage{
load_model(path)
}
\arguments{
\item{path}{the path that stores the model, make sure ended with .rds}
}
\value{
the model object
}
\description{
Load the model
}
\examples{
mod <- load_model(paste0(getwd(), "theModel.rds"))
}
\author{
Shangchen Song
}
