% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NACCADNI-package.R
\docType{package}
\name{NACCADNI-package}
\alias{NACCADNI}
\alias{NACCADNI-package}
\title{NACCADNI: Predicting Dementia due to AD Survival Probability}
\description{
Delivering prediction models for NACC and ADNI registered cohort
}
\author{
\strong{Maintainer}: Shangchen Song \email{s.song@ufl.edu} (\href{https://orcid.org/0000-0001-5145-7404}{ORCID})

}
\keyword{internal}
