% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{plot_survProb}
\alias{plot_survProb}
\title{Plot the survival curves}
\usage{
plot_survProb(survProb.df)
}
\arguments{
\item{survProb.df}{the survival probability data.frame output by \code{survProb_predict}}
}
\value{
No return value
}
\description{
Plot the survival curves
}
\examples{
plot_survProb(survProb_predict(test.pred))
}
\author{
Shangchen Song
}
