% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{predict_model}
\alias{predict_model}
\title{Make prediction on the new dataset}
\usage{
predict_model(model, newdata)
}
\arguments{
\item{model}{the loaded model}

\item{newdata}{data.frame that has six variables}
}
\value{
predicted object
}
\description{
Make prediction on the new dataset
}
\examples{
test.pred <- predict_model(mod, dat)
}
\author{
Shangchen Song
}
