% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{survProb_predict}
\alias{survProb_predict}
\title{Extract the survival probability table}
\usage{
survProb_predict(predict.out)
}
\arguments{
\item{predict.out}{the predicted object}
}
\value{
data.frame
}
\description{
Extract the survival probability table
}
\examples{
survProb_predict(test.pred)
}
\author{
Shangchen Song
}
