% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinydat.R
\docType{data}
\name{tinydat}
\alias{tinydat}
\title{A fake example dataset}
\format{
A data frame with 10 rows and 6 variables:
\describe{
\item{MEMUNITS}
\item{CDRSUM}
\item{REMDATES}
\item{BILLS}
\item{ORIENT}
\item{NACCAGE}
...
}
}
\source{
fake data
}
\usage{
tinydat
}
\description{
A dataset containing six variables used in our models
}
\keyword{datasets}
